+++ lib/rodauth/rails/app.rb
@@ -11,6 +11,7 @@ module Rodauth
 
       plugin :hooks
       plugin :render, layout: false
+      plugin :pass
 
       unless Rodauth::Rails.api_only?
         require "rodauth/rails/app/flash"